third_party/blink/web_tests/external/wpt/css/css-masking/parsing/mask-origin-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-origin with invalid values</title>
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-origin">
<meta name="assert" content="mask-origin supports only the grammar '<coord-box>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-origin", "margin-box");
test_invalid_value("mask-origin", "no-clip");
</script>
</body>
</html>
